generalised apiPdf download
