+++ src/logic/patientApi.ts
@@ -4,7 +4,7 @@ import { EditPatientDetailsData } from "../types/patientAPITypes";
 import { Patient, Role } from "../types/users";
 import { BASE_URL } from './config';
 import { Timeslot, Vaccine, validDiseases } from "../types/vaccination";
-import { apiGet, apiPost, apiPut, checkStatusAndGetBody, checkStatusAndIgnoreBody } from "./API";
+import {apiGet, apiGetPdf, apiPost, apiPut, checkStatusAndGetBody, checkStatusAndIgnoreBody} from "./API";
 import {StatusCodes} from "http-status-codes";
 import {UnauthorizedRequestError} from "../types/requestErrors";
 
@@ -15,18 +15,7 @@ export function registerPatient(registrationData: RegistrationData) {
 }
 
 export function downloadCertificate(auth: AuthContextType, id: number) {
-    return fetch(`${BASE_URL}/patient/vaccinations/${id}/certificate`, {
-        method: 'GET',
-        headers: {
-            'Content-Type': 'application/pdf',
-            'Authorization': auth?.token ?? ''
-        }
-    }).then((response) => {
-        if (response.ok) return response.blob();
-        if (response.status === StatusCodes.UNAUTHORIZED)
-            throw new UnauthorizedRequestError('You are not authorized');
-        throw response;
-    }).then(showFile)
+    return apiGetPdf(`${BASE_URL}/patient/vaccinations/${id}/certificate`, auth)
 }
 
 export function editPatientDetails(auth: AuthContextType, 
@@ -66,22 +55,4 @@ export function reserveTimeslot(auth: AuthContextType, timeslot: Timeslot,
     return apiPut(`${BASE_URL}/patient/vaccination-slots/${timeslot.id}`, auth,
             `{"vaccineId": ${vaccine.id}}`)
         .then(checkStatusAndIgnoreBody);
-}
-
-function showFile(blob: BlobPart){
-    // It is necessary to create a new blob object with mime-type explicitly set
-    // otherwise only Chrome works like it should
-    const newBlob = new Blob([blob], {type: "application/pdf"});
-
-    // For other browsers:
-    // Create a link pointing to the ObjectURL containing the blob.
-    const data = window.URL.createObjectURL(newBlob);
-    const link = document.createElement('a');
-    link.href = data;
-    link.download="file.pdf";
-    link.click();
-    setTimeout(function(){
-        // For Firefox it is necessary to delay revoking the ObjectURL
-        window.URL.revokeObjectURL(data);
-    }, 100);
 }
